Refactor to move alloc stuff to corealloc.c (1/8)
This series of commits refactors the kernel scheduler to move all core
allocation logic out into a file called corealloc.c. In this first
commit, we move the idlecore list and the all_pcores list over to this
file along with their initialization. Future commits will move the rest
of the logic over.
